% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govuk-format.R
\name{govuk_document}
\alias{govuk_document}
\title{GOV.UK style HTML template}
\usage{
govuk_document(...)
}
\arguments{
\item{...}{additional arguments provided to \code{rmarkdown::html_document}}
}
\description{
Loads additional style and template file
}
