% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govuk-format.R
\name{govuk_document}
\alias{govuk_document}
\title{GOV.UK style HTML template}
\usage{
govuk_document(..., phase = c("none", "alpha", "beta"),
  service_name = NULL, css = NULL, extra_dependencies = NULL,
  pandoc_args = NULL)
}
\arguments{
\item{...}{Additional function arguments to pass to the
base R Markdown HTML output formatter \code{\link{html_document_base}}}

\item{css}{One or more css files to include}

\item{extra_dependencies}{Additional function arguments to pass to the
base R Markdown HTML output formatter \code{\link{html_document_base}}}

\item{pandoc_args}{Additional command line options to pass to pandoc}
}
\description{
Loads additional style and template file
}
